refactor(web): isolate integration capability contracts
Keep FactoryIntegration focused on lifecycle composition while Intake and VersionControl evolve independently in dedicated modules.\n\nCo-Authored-By: Mastra Code (openai/gpt-5.6-sol) <noreply@mastra.ai>
