+++ mastracode/web/src/web/capabilities/connection.ts
@@ -1,0 +1,3 @@
+/** Authentication material accepted by integration capabilities. */
+export type IntegrationConnection =
+  { type: 'app-installation'; installationId: number } | { type: 'oauth'; accessToken: string };
