+++ mastracode/web/src/web/github/routes-scenario.test.ts
@@ -100,7 +100,7 @@ const githubStub = {
   // A fresh token string per call so the scenario can prove per-op minting.
   mintInstallationToken: vi.fn(async () => `install-token-${++mintCount}`),
   versionControl: {
-    createPullRequest: async (input: import('../factory-integration').CreatePullRequestInput) => {
+    createPullRequest: async (input: import('../capabilities/version-control').CreatePullRequestInput) => {
       const result = await createPullRequest(input);
       return {
         id: '1',
